cluster_STspot: Detect clusters on ST arrays

Perform unsupervised spatially-informed clustering on the spots of an spatial array.

cluster_STspot(
  x = NULL,
  weights = 0.1,
  pcs = NULL,
  vperc = NULL,
  method = "ward.D",
  ks = "dtc",
  spotfilter = 0,
  topgenes = 2000
)

Arguments

x
an STList with normalized expression data.

weights
a double 0-1 indicating the weight to be applied to spatial distances. Defaults to w=0.1.

pcs
the number of principal components (PCs) to retain.

vperc
the minimum percentage of explained variance explained by the PCs to decide how many of them to retain.

method
the linkage method applied to hierarchical clustering. It is passed to hclust and defaults to 'ward.D'.

ks
the range of k values to assess. Defaults to dtc, meaning cutreeDynamic is applied.

spotfilter
the number of genes with more than zero counts for a spot to be included in the clustering analysis. The lower this number, the slower the analysis. If spotfilter=0, all spots are retained.

topgenes
the number of high spot-to-spot standard deviation to retain before PCA.

Value

x, the STList with cluster assignments.

Details

The function takes an STList and performs PCA on the gene expression data, then calculates euclidean distances of the PCs and spatial distances. Weighted averages between the two distances are calculated to perform hierarchical clustering. The user can define how much weight the spatial distance should have (values between 0.05 to 0.25 work reasonably well). The method aims to detect compartments within a tissue.

Examples

# In this example, melanoma is an STList.
# Using Dynamic Tree Cuts:
melanoma <- cluster_STspot(melanoma, ks='dtc', weights=0.1)

#> Loading required package: magrittr

#> Error in cluster_STspot(melanoma, ks = "dtc", weights = 0.1): object 'melanoma' not found

# Using a range of ks:
melanoma <- cluster_STspot(melanoma, ks=c(2:6), weights=0.1)

#> Error in cluster_STspot(melanoma, ks = c(2:6), weights = 0.1): object 'melanoma' not found