filter_data: Removes spots, genes, or samples

Provides methods to remove spots and genes or entire samples.

filter_data(
  x = NULL,
  spot_minreads = NULL,
  spot_maxreads = NULL,
  spot_mingenes = NULL,
  spot_maxgenes = NULL,
  spot_minpct = NULL,
  spot_maxpct = NULL,
  gene_minreads = NULL,
  gene_maxreads = NULL,
  gene_minspots = NULL,
  gene_maxspots = NULL,
  gene_minpct = NULL,
  gene_maxpct = NULL,
  who = NULL,
  rm_tissue = NULL,
  rm_genes = NULL,
  rm_partialgenes = NULL,
  spot_pctExpr = "^MT-"
)

Arguments

x
an STList

spot_minreads
@param spot_maxreads the minimum (or maximum) number of total reads for a spot to be retained

spot_mingenes
@param spot_maxgenes the minimum (or maximum) number of non-zero counts for a spot to be retained

spot_minpct
@param spot_maxpctExpr the minimum (or maximum) percentage of counts for features defined by spot_pctexpr for a spot to be retained. By default, mtDNA genes.

gene_minreads
@param gene_maxreads the minimum (or maximum) number of total reads for a gene to be retained

gene_minspots
@param gene_maxspots the minimum (or maximum) number of spots with non-zero counts for a gene to be retained

gene_minpct
@param gene_maxpct the minimum (or maximum) percentage of spots with non-zero counts for a gene to be retained

who
vector of integers (as in x@counts) or sample names to perform filtering

rm_tissue
vector of integers (as in x@counts) or sample names to remove from STList

rm_genes
vector of gene names to remove from STList

rm_partialgenes
a expression to match gene names. Useful to remove entire gene classes (for example mtDNA '^MT-')

spot_pctExpr
a expression to use with spot_minpct. By default '^MT-'.

Details

This function provides methods to filter an STList. It can remove spots or genes based on absolute counts or percentages. Users can input a expression to be matched with gene names to calculate percentages (for example %mtDNA genes). Note that the function removes counts raw counts and spatial information. Transformed counts or other results within the STList remain unaffected.