plot_STclusters: Plot cluster memberships of ST spots

Generates a plot of the location of spots within an spatial array, and colors them according to spatially-weighted hierarchical clustering assignments.

plot_STclusters(
  x,
  plot_who = NULL,
  ks = NULL,
  weights = NULL,
  purity = F,
  color_pal = "light",
  visium = T,
  ptsize = NULL
)

Arguments

x

an STList with hierarchical cluster memberships.

plot_who

an integer indicating the spatial array(s) to be plotted. Numbers follow the order of names(x@counts). If NULL, will plot all spatial arrays.

ks

the k values to plot

weights

the spatial weights to plot

purity

logical, whether or not annotate tumor/stroma.

color_pal

a string of a color palette from khroma or RColorBrewer, or a vector with colors with enough elements to plot categories.

visium

whether or not to reverse axes for Visium slides.

ptsize,

a number specifying the size of the points. Passed to size aesthetic.

Value

a list with the requested plots.

Details

The function takes an STList with cluster memberships and plots the spots with colors indicating the cluster they belong to. Optionally, the user can annotate tumor/stroma compartments if ESTIMATE scores are available.

Examples

# In this example, melanoma is an STList.
# cluster_p <- plot_STclusters(melanoma, purity=T, plot_who=c(2,3), visium=F)