quilt_p

Creates a quilt plot from ST data.

quilt_p(
  data_f = NULL,
  color_pal = "YlOrBr",
  leg_name = "",
  title_name = "",
  minvalue = minvalue,
  maxvalue = maxvalue,
  visium = T,
  ptsize = 0.5
)

Arguments

data_f,

a data with three columns: x coordinates, y coordinates, and the values to be plotted.

color_pal,

a scheme from 'khroma'. Default is 'YlOrBr'.

leg_name,

a short name for the legend title.

title_name,

a short name for the plot title.

minvalue,

the minimum value of gene expression or cell score. Used for standardization.

maxvalue,

the maximum value of gene expression or cell score. Used for standardization.

visium,

whether or not to reverse axes for Visium slides.

ptsize,

a number specifying the size of the points. Passed to size aesthetic.

Value

, a ggplot object.

Details

Function to produce a "quilt plot" from a data frame with three columns: x coordinates, y coordinates, and values (expression or cell scores). The data frame has column names 'x_pos', 'y_pos', and 'values'. It also takes a color palette name from the 'khroma' or RColorBrewer packages. Finally, it takes a name for the color legend title.