STList: Creation of STList objects

Creates an STList object from one or several spatial transcriptomic samples.

Usage

STList(
  rnacounts = NULL,
  spotcoords = NULL,
  samples = NULL,
  gmx_pkc = NULL,
  gmx_slide_col = NULL,
  gmx_roi_col = NULL,
  gmx_x_col = NULL,
  gmx_y_col = NULL,
  gmx_meta_cols = NULL
)

Arguments

rnacounts

The count data which can be provided in one of these formats:

File paths to comma- or tab-delimited files containing raw RNA counts, one for each spatial array. The first column contains gene names. Subsequent columns contain data for each spot. Duplicate gene names will be appended a suffix ('_d**'). Requires spotcoords and samples.

File paths to Visium output folders (one per spatial array). The folder should have the structure resulting from spaceranger count and contain the .h5 files and spatial folder. Requires samples.

File path to .dcc files from GeoMx output. Requires samples.

One named list of data frames with raw RNA counts (one data frame per spatial array). Requires spotcoords and samples.

spotcoords

The spot coordinates, which are required if inputs are not Visium outputs. The user can provide one of these formats:

File paths to comma- or tab-delimited files containing spot coordinates, one for each spatial array. The files must contain three columns: spot IDs, y positions, and x positions. The spot IDs must match the column names for each spot (column) in the RNA count files.

One named list of data frames with spot coordinates. The list names must match list names of the RNA counts list.

samples

The metadata associated to each spatial array. This file can also include system paths to RNA count and spot coordinate files, bypassing the need to specify rnacounts and spotcoords. One of the following options should be entered to create an STList:

A vector with sample names, which will be used to partial match RNA count and spot coordinates file paths. A sample name must not match two file paths.

A path to the file containing metadata for each spatial array. This file is a comma- or tab-separated file with one sample per row and sample names in the first column. Paths to RNA count and coordinate files can be in the second and third column respectively, and omitting the rnacounts and spotcoords arguments. If Visium, only the second column with paths to output folders is expected. Subsequent columns can contain variables associated to each spatial array. Note: For GeoMx, the metadata file contains one row per ROI. This information is later summarized to one row per slide by this function.

gmx_pkc,

the file path to the .pkc (for GeoMx input)

gmx_slide_col,

the name of the column in the metadata table containing the slide names (for GeoMx input)

gmx_roi_col,

the name of the column in the metadata table containing the ROI IDs, matching IDs in the DCC files (for GeoMx input)

gmx_x_col,

the name of the column in the metadata table containing the x coordinates (for GeoMx input)

gmx_y_col,

the name of the column in the metadata table containing the y coordinates (for GeoMx input)

gmx_meta_cols,

a vector with column names in the metadata table containing clinical data (for GeoMx input)

Value

x, the STList object containing the counts and coordinates, and optionally the sample metadata.

Details

Objects of the S4 class STList can be created from two sources:

Raw RNA counts (genes in rows, spots in columns) and spot coordinates (spots in rows, with three columns: spot IDs, y position, and x position).

Visium output folders from Space Ranger. The folder should have the structure resulting from spaceranger count.

Optionally, the user can input a system path to a table containing data associated with each spatial array (e.g., clinical variables). This sample metadata file should contain sample IDs in the first column matching partially the file names of the count/coordinate file paths or Visium directories.

The function will read data in parallel if unix system is available (Windows users will experience longer times depending on the number of samples).

Examples

# Using included melanoma example (Thrane et al.)
# data_files <- system.file("extdata", package="spatialGEdev")
# count_files <- grep("genes", data_files, value=T)
# coord_files <- grep("mapping", data_files, value=T)
# clin_file <- grep("clinical", data_files, value=T)
# melanoma <- STList(rnacounts=count_files, spotcoords=coord_files, samples=clin_file)
# melanoma