STplot_interpolation: Visualize gene expression surfaces

Produces a gene expression surface from kriging interpolation of ST data.

STplot_interpolation(
  x = NULL,
  genes = NULL,
  top_n = 10,
  samples = NULL,
  color_pal = "BuRd"
)

Arguments

x

an STlist containing results from gene_krige for the genes selected.

genes

a vector of gene names (one or several) to plot. If 'top', the 10 genes with highest standard deviation from each spatial sample are plotted.

top_n

an integer indicating how many top genes to perform kriging. Default is 10.

samples

a vector indicating the spatial samples to plot. If vector of numbers, it follows the order of names(x@counts). If NULL, the function plots all samples

color_pal

a color scheme from khroma or RColorBrewer.

Value

a list of plots

Details

This function produces a gene expression surface plot via kriging for one or several genes and spatial samples

Examples

# Using included melanoma example (Thrane et al.)
library('spatialGE')
data_files <- list.files(system.file("extdata", package="spatialGE"), recursive=T, full.names=T)
count_files <- grep("counts", data_files, value=T)
coord_files <- grep("mapping", data_files, value=T)
clin_file <- grep("thrane_clinical", data_files, value=T)
melanoma <- STlist(rnacounts=count_files[c(1,2)], spotcoords=coord_files[c(1,2)], samples=clin_file) # Only first two samples
#> Warning: Sample ST_mel2_rep1 was not found among the count/coordinate files.
#> Warning: Sample ST_mel2_rep2 was not found among the count/coordinate files.
#> Warning: Sample ST_mel3_rep1 was not found among the count/coordinate files.
#> Warning: Sample ST_mel3_rep2 was not found among the count/coordinate files.
#> Warning: Sample ST_mel4_rep1 was not found among the count/coordinate files.
#> Warning: Sample ST_mel4_rep2 was not found among the count/coordinate files.
#> Found matrix data
#> Matching gene expression and coordinate data...
#> Converting counts to sparse matrices
#> Completed STlist!
#> 
melanoma <- transform_data(melanoma)
melanoma <- gene_interpolation(melanoma, genes=c('MLANA', 'COL1A1'), samples='ST_mel1_rep2')
#> Gene interpolation started.
#> Gene interpolation completed in 0.03 min.
#> 
kp = STplot_interpolation(melanoma, genes=c('MLANA', 'COL1A1'))
#> Error in if (x@gene_krige[[gene]][[i]][["success_or_not"]] != "error") {    values <- x@gene_krige[[gene]][[i]][["krige_out"]][["z"]]    maxvalue <- append(maxvalue, max(values, na.rm = T))    minvalue <- append(minvalue, min(values, na.rm = T))}: argument is of length zero
ggpubr::ggarrange(plotlist=kp)
#> Error in ggpubr::ggarrange(plotlist = kp): object 'kp' not found